Renamed dictionaries to high-score-board to reflect the reorg from concepts -> exercises Added missing entry in config.json Added dictionaries to concepts along with about.md and links.json
