--- languages/swift/exercises/concept/dictionaries/Tests/LinuxMain.swift
@@ -1,6 +1,0 @@
-import XCTest
-import dictionariesTests
-
-var tests = [XCTestCaseEntry]()
-tests += dictionariesTests.allTests()
-XCTMain(tests)
