+++ languages/swift/exercises/concept/high-score-board/Tests/HighScoreBoardTests/XCTestManifests.swift
@@ -3,7 +3,7 @@ import XCTest
 #if !canImport(ObjectiveC)
   public func allTests() -> [XCTestCaseEntry] {
     return [
-      testCase(dictionariesTests.allTests)
+      testCase(HighScoreBoardTests.allTests)
     ]
   }
 #endif
