+++ languages/swift/exercises/concept/high-score-board/Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import HighScoreBoardTests
+import XCTest
+
+var tests = [XCTestCaseEntry]()
+tests += HighScoreBoardTests.allTests()
+XCTMain(tests)
